More precise CPU clock calculation

+++ src/cpu/6502.h
@@ -50,10 +50,10 @@ extern const double default_MHz_6502;
 extern const double iigs_MHz_6502;
 extern const double iicplus_MHz_6502;
 extern double MHz_6502;
-extern unsigned long long clk_6502_per_frm;
-extern unsigned long long clk_6502_per_frm_set;
-extern unsigned long long clk_6502_per_frm_max;
-extern unsigned long long clk_6502_per_frm_max_sound;
+extern unsigned int clk_6502_per_frm;
+extern unsigned int clk_6502_per_frm_set;
+extern unsigned int clk_6502_per_frm_max;
+extern unsigned int clk_6502_per_frm_max_sound;
 
 
 typedef enum {
